Error when Windows nrnivmodl or RxD cannot find cl.exe.
The pip wheel does not ship a compiler. Fail before cmake with the
Build Tools URL instead of a missing CMAKE_CXX_COMPILER. setup.exe
MinGW g++ is unchanged. Same vswhere lookup as RxD, shared in
neuron._windows_cxx.
