+++ share/lib/python/scripts/binwrapper.py
@@ -194,8 +194,21 @@ def _nrnivmodl_cmake(args):
     cmake = shutil.which("cmake")
     if not cmake:
         raise SystemExit(
-            "nrnivmodl: cmake not found on PATH; install CMake and a C++ compiler"
+            "nrnivmodl: cmake not found on PATH; install CMake from https://cmake.org/download/"
         )
+
+    if os.name == "nt":
+        from neuron._windows_cxx import MSVC_CXX_MISSING, msvc_cl_available
+
+        mingw = (
+            Path(os.environ.get("NRNHOME", ""))
+            / "mingw"
+            / "mingw64"
+            / "bin"
+            / "x86_64-w64-mingw32-g++.exe"
+        )
+        if not mingw.is_file() and not msvc_cl_available():
+            raise SystemExit(MSVC_CXX_MISSING.strip())
 
     prefix = Path(os.environ["NRNHOME"])
     srcdir = prefix / "lib" / "cmake" / "neuron" / "nrnivmodl"
